Warnings fix

+++ Bermuda/Bermuda/GameTimer.cpp
@@ -34,8 +34,9 @@ void GameTimer::init()
 
 void GameTimer::updateGameTime(double frameTime)
 {
+
 	this->frameTime = frameTime;
-	this->gameTime += frameTime;
+	this->gameTime += static_cast<float>(frameTime);
 	this->updateDay();
 }
